# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /magic:
    post:
      tags:
        - magic-8-ball-api-controller
      operationId: getAnswer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicQuestion'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Answer'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /word:
    get:
      tags:
        - word-api-controller
      operationId: getWord
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Definition'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /quote:
    get:
      tags:
        - quote-api-controller
      operationId: getQuote
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Quote'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
        statusCode:
          type: integer
          format: int32
    MagicQuestion:
      required:
        - text
      type: object
      properties:
        text:
          type: string
    Answer:
      type: object
      properties:
        id:
          type: integer
          format: int32
        question:
          $ref: '#/components/schemas/MagicQuestion'
        answer:
          type: string
    Definition:
      type: object
      properties:
        id:
          type: integer
          format: int32
        word:
          type: string
        definition:
          type: string
    Quote:
      type: object
      properties:
        id:
          type: integer
          format: int32
        author:
          type: string
        quote:
          type: string