# src/main/java/com/company/Summative1AssessmentAguadzeMichael/models/Definition.java
package com.company.Summative1AssessmentAguadzeMichael.models;

public class Definition {

    private int id;

    private String word;

    private String definition;

    public int getId() {
        return id;
    }

    public Definition(int id, String word, String definition) {
        this.id = id;
        this.word = word;
        this.definition = definition;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getWord() {
        return word;
    }

    public void setWord(String word) {
        this.word = word;
    }

    public String getDefinition() {
        return definition;
    }

    public void setDefinition(String definition) {
        this.definition = definition;
    }
}